Refatorando comunicação
